chg: dev: Code refactoring
