+++ src/modules/view-components/background-animation-view.ts
